+++ alacritty/res/glsl3/text.f.glsl
@@ -1,4 +1,3 @@
-#version 330 core
 in vec2 TexCoords;
 flat in vec4 fg;
 flat in vec4 bg;
