+++ alacritty/res/glsl3/text.v.glsl
@@ -1,4 +1,3 @@
-#version 330 core
 // Cell properties.
 layout(location = 0) in vec2 gridCoords;
