+++ alacritty/res/rect.f.glsl
@@ -1,18 +1,32 @@
-#version 330 core
+#if defined(GLES2_RENDERER)
+#define float_t mediump float
+#define color_t mediump vec4
+#define FRAG_COLOR gl_FragColor
 
-flat in vec4 color;
+varying color_t color;
+
+#else
+#define float_t float
+#define int_t int
+#define color_t vec4
+
 out vec4 FragColor;
+#define FRAG_COLOR FragColor
+
+flat in color_t color;
+
+#endif
+
 uniform int rectKind;
+uniform float_t cellWidth;
+uniform float_t cellHeight;
+uniform float_t paddingY;
+uniform float_t paddingX;
 
-uniform float cellWidth;
-uniform float cellHeight;
-uniform float paddingY;
-uniform float paddingX;
+uniform float_t underlinePosition;
+uniform float_t underlineThickness;
 
-uniform float underlinePosition;
-uniform float underlineThickness;
-
-uniform float undercurlPosition;
+uniform float_t undercurlPosition;
 
 #define UNDERCURL 1
 #define DOTTED 2
@@ -20,19 +34,18 @@ uniform float undercurlPosition;
 
 #define PI 3.1415926538
 
-vec4 draw_undercurl(int x, int y) {
+color_t draw_undercurl(float_t x, float_t y) {
   // We use `undercurlPosition` as an amplitude, since it's half of the descent
   // value.
-  float undercurl =
-      undercurlPosition / 2. * cos(float(x) * 2 * PI / cellWidth) +
-      + undercurlPosition - 1.;
+  float_t undercurl = undercurlPosition / 2. * cos(x * 2. * PI / cellWidth)
+                    + undercurlPosition - 1.;
 
-  float undercurlTop = undercurl + max((underlineThickness - 1), 0);
-  float undercurlBottom = undercurl - max((underlineThickness - 1), 0);
+  float_t undercurlTop = undercurl + max((underlineThickness - 1.), 0.);
+  float_t undercurlBottom = undercurl - max((underlineThickness - 1.), 0.);
 
   // Compute resulted alpha based on distance from `gl_FragCoord.y` to the
   // cosine curve.
-  float alpha = 1.;
+  float_t alpha = 1.;
   if (y > undercurlTop || y < undercurlBottom) {
     alpha = 1. - min(abs(undercurlTop - y), abs(undercurlBottom - y));
   }
@@ -43,54 +56,54 @@ vec4 draw_undercurl(int x, int y) {
 
 // When the dot size increases we can use AA to make spacing look even and the
 // dots rounded.
-vec4 draw_dotted_aliased(float x, float y) {
-  int dotNumber = int(x / underlineThickness);
+color_t draw_dotted_aliased(float_t x, float_t y) {
+  float_t dotNumber = floor(x / underlineThickness);
 
-  float radius = underlineThickness / 2.;
-  float centerY = underlinePosition - 1.;
+  float_t radius = underlineThickness / 2.;
+  float_t centerY = underlinePosition - 1.;
 
-  float leftCenter = (dotNumber - dotNumber % 2) * underlineThickness + radius;
-  float rightCenter = leftCenter + 2 * underlineThickness;
+  float_t leftCenter = (dotNumber - mod(dotNumber, 2.)) * underlineThickness + radius;
+  float_t rightCenter = leftCenter + 2. * underlineThickness;
 
-  float distanceLeft = sqrt(pow(x - leftCenter, 2) + pow(y - centerY, 2));
-  float distanceRight = sqrt(pow(x - rightCenter, 2) + pow(y - centerY, 2));
+  float_t distanceLeft = sqrt(pow(x - leftCenter, 2.) + pow(y - centerY, 2.));
+  float_t distanceRight = sqrt(pow(x - rightCenter, 2.) + pow(y - centerY, 2.));
 
-  float alpha = max(1 - (min(distanceLeft, distanceRight) - radius), 0);
+  float_t alpha = max(1. - (min(distanceLeft, distanceRight) - radius), 0.);
   return vec4(color.rgb, alpha);
 }
 
 /// Draw dotted line when dot is just a single pixel.
-vec4 draw_dotted(int x, int y) {
-  int cellEven = 0;
+color_t draw_dotted(float_t x, float_t y) {
+  float_t cellEven = 0.;
 
   // Since the size of the dot and its gap combined is 2px we should ensure that
   // spacing will be even. If the cellWidth is even it'll work since we start
   // with dot and end with gap. However if cellWidth is odd, the cell will start
   // and end with a dot, creating a dash. To resolve this issue, we invert the
   // pattern every two cells.
-  if (int(cellWidth) % 2 != 0) {
-    cellEven = int((gl_FragCoord.x - paddingX) / cellWidth) % 2;
+  if (int(mod(cellWidth, 2.)) != 0) {
+    cellEven = mod((gl_FragCoord.x - paddingX) / cellWidth, 2.);
   }
 
   // Since we use the entire descent area for dotted underlines, we limit its
   // height to a single pixel so we don't draw bars instead of dots.
-  float alpha = 1. - abs(round(underlinePosition - 1.) - y);
-  if (x % 2 != cellEven) {
-    alpha = 0;
+  float_t alpha = 1. - abs(floor(underlinePosition - 0.5) - y);
+  if (int(mod(x, 2.)) != int(cellEven)) {
+    alpha = 0.;
   }
 
   return vec4(color.rgb, alpha);
 }
 
-vec4 draw_dashed(int x) {
+color_t draw_dashed(float_t x) {
   // Since dashes of adjacent cells connect with each other our dash length is
   // half of the desired total length.
-  int halfDashLen = int(cellWidth) / 4;
+  float_t halfDashLen = floor(cellWidth / 4.);
 
-  float alpha = 1.;
+  float_t alpha = 1.;
 
   // Check if `x` coordinate is where we should draw gap.
-  if (x > halfDashLen && x < cellWidth - halfDashLen - 1) {
+  if (x > halfDashLen && x < cellWidth - halfDashLen - 1.) {
     alpha = 0.;
   }
 
@@ -98,25 +111,20 @@ vec4 draw_dashed(int x) {
 }
 
 void main() {
-  int x = int(gl_FragCoord.x - paddingX) % int(cellWidth);
-  int y = int(gl_FragCoord.y - paddingY) % int(cellHeight);
+  float_t x = floor(mod(gl_FragCoord.x - paddingX, cellWidth));
+  float_t y = floor(mod(gl_FragCoord.y - paddingY, cellHeight));
 
-  switch (rectKind) {
-  case UNDERCURL:
-    FragColor = draw_undercurl(x, y);
-    break;
-  case DOTTED:
-    if (underlineThickness < 2) {
-      FragColor = draw_dotted(x, y);
+  if (rectKind == UNDERCURL) {
+    FRAG_COLOR = draw_undercurl(x, y);
+  } else if (rectKind == DOTTED) {
+    if (underlineThickness < 2.) {
+      FRAG_COLOR = draw_dotted(x, y);
     } else {
-      FragColor = draw_dotted_aliased(x, y);
+      FRAG_COLOR = draw_dotted_aliased(x, y);
     }
-    break;
-  case DASHED:
-    FragColor = draw_dashed(x);
-    break;
-  default:
-    FragColor = color;
-    break;
+  } else if (rectKind == DASHED) {
+    FRAG_COLOR = draw_dashed(x);
+  } else {
+    FRAG_COLOR = color;
   }
 }
